nlptosqlagent/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
#sqlalchemy==2.1.0
sqlalchemy
pyodbc==4.0.35
pandas==2.2.2
langchain_openai==0.1.8
#langchain_core==0.1.22
langchain_core
python-multipart==0.0.6

